Modified the dump method of PrintSeries so that it dumps the series values at each iteration

+++ Homework2/src/printSeries.hh
@@ -6,7 +6,7 @@
 
 class PrintSeries: public DumperSeries{
 public:
-    PrintSeries(Series & series, double frequency, int maxiter);
+    PrintSeries(Series & series, int frequency, int maxiter);
     void dump() override;
 private:
     double frequency;
